review: neither engine class imports the other; the delta map is exhaustive; tool projection tested
- Shared turn plumbing (lease, terminal classification, image conversion, the
  push→pull fan-in) moves to turn-plumbing.ts. session-invoke.ts was importing the
  harness L0, which made the choice of engine class also a choice of which module
  graph loads; now both siblings import the plumbing and neither imports the other.
- The delta guard pinned two names, so pi ADDING a channel — the drift that actually
  goes silent — would pass. Replaced with an exhaustive project/ignore map over
  AssistantMessageEvent['type'], verified to fail typecheck when a member is dropped.
- tool_started/tool_ended carried untested payloads. The new test also corrected a
  fixture written from assumption: pi's AgentToolResult has no isError field, a tool
  reports failure by throwing.
