+++ src/engines/pi/invoke.ts
@@ -13,7 +13,7 @@
  */
 import type { AgentHarnessEvent } from "@earendil-works/pi-agent-core";
 import { DEFAULT_COMPACTION_SETTINGS, calculateContextTokens, shouldCompact } from "@earendil-works/pi-agent-core";
-import type { AssistantMessage, ImageContent } from "@earendil-works/pi-ai";
+import type { AssistantMessage } from "@earendil-works/pi-ai";
 import {
   ABORTED_CODE,
   type Agent,
@@ -26,38 +26,16 @@ import {
 import { abortFirstIterator } from "../../collect.ts";
 import type { RetryScheduledEvent, RunSettledEvent, SessionEvent } from "../../session.ts";
 import { log } from "../../log.ts";
+import {
+  EventQueue,
+  type Lease,
+  errorToTerminal,
+  inProcessLease,
+  toPiPromptOptions,
+  toTerminal,
+} from "./turn-plumbing.ts";
 import { TOOL_ACTIVATION_ENTRY, harnessSession, type PiHarnessFactory } from "./harness.ts";
 import { type ReadonlySessionManager, type ToolActivation, additiveActivation, turnContext } from "./tool-context.ts";
-
-// ── §1 Lease: single-writer concurrency floor ───────────────────────────────
-//
-// Corruption-prevention floor only: it does not pick a UX. Fail-fast over queueing because real
-// same-session concurrency is mostly duplicate intent or a user firing follow-ups, not two real
-// turns; a queue would also leak a slot when a waiter is cancelled. Synchronous (no awaits) so
-// nothing interleaves between acquire and entering try — cancellation always releases in finally.
-
-export type Release = () => void;
-
-export interface Lease {
-  /** Try to acquire exclusive write access for the session (fail-fast). Returns null if held. */
-  tryAcquire(session: string): Release | null;
-}
-
-export function inProcessLease(): Lease {
-  const busy = new Set<string>();
-  return {
-    tryAcquire(session: string): Release | null {
-      if (busy.has(session)) return null;
-      busy.add(session);
-      let released = false;
-      return () => {
-        if (released) return;
-        released = true;
-        busy.delete(session);
-      };
-    },
-  };
-}
 
 // ── §2 translate: the single pi↔SPEC translation point ───────────────────────
 //
@@ -75,67 +53,6 @@ export function inProcessLease(): Lease {
 // already exhausted the cleanly-retryable cases. The regex is the narrow ceiling, not the classifier.
 // Upstream ask: a first-class `retryable`/`kind` on pi's terminal error would retire the prose path
 // entirely (mirrors the §11 "the deeper fix is upstream in pi" pattern).
-
-/** Clearly-transient network error codes (Node/undici), decisive on their own. */
-const RETRYABLE_CODES = new Set([
-  "ECONNRESET",
-  "ETIMEDOUT",
-  "ENETUNREACH",
-  "ENETDOWN",
-  "EAI_AGAIN",
-  "EPIPE",
-  "UND_ERR_CONNECT_TIMEOUT",
-  "UND_ERR_SOCKET",
-]);
-
-/** 429 (rate limit) and 5xx (server) are worth retrying; other statuses are decisive NON-retryable. */
-const statusIsRetryable = (status: number): boolean => status === 429 || (status >= 500 && status < 600);
-
-/** Last-resort prose match, used only when no structured status/code is available. */
-const RETRYABLE_MESSAGE =
-  /\b(429|5\d\d|timeout|timed out|rate.?limit|overloaded|ECONNRESET|ETIMEDOUT|ENETUNREACH|EAI_AGAIN|socket hang up)\b/i;
-
-/** A structured status/code decision, or `null` when the signal is absent/undecisive → fall to prose. */
-function retryableFromSignal(signal: { status?: number; code?: unknown }): boolean | null {
-  if (typeof signal.status === "number") return statusIsRetryable(signal.status);
-  const { code } = signal;
-  if (typeof code === "number") return statusIsRetryable(code);
-  if (typeof code === "string") {
-    if (RETRYABLE_CODES.has(code)) return true;
-    if (/^\d{3}$/.test(code)) return statusIsRetryable(Number(code)); // a status carried as a string
-  }
-  return null; // no code, or an unknown one — not decisive on its own
-}
-
-/** Classify `retryable`: structured status/code first, message prose only as the last-resort ceiling. */
-export function classifyRetryable(details: string, signal: { status?: number; code?: unknown }): boolean {
-  return retryableFromSignal(signal) ?? RETRYABLE_MESSAGE.test(details);
-}
-
-/** Pull a structured status/code off a thrown error (HTTP status or a network code, incl. its cause). */
-function errorSignal(error: unknown): { status?: number; code?: unknown } {
-  if (!error || typeof error !== "object") return {};
-  const e = error as { status?: unknown; statusCode?: unknown; code?: unknown; cause?: unknown };
-  const status = typeof e.status === "number" ? e.status : typeof e.statusCode === "number" ? e.statusCode : undefined;
-  const causeCode = e.cause && typeof e.cause === "object" ? (e.cause as { code?: unknown }).code : undefined;
-  return { status, code: e.code ?? causeCode };
-}
-
-/**
- * Pull the structured error `code` pi records on a failed message's diagnostics. `diagnostics`
- * accumulates across attempts (`appendAssistantMessageDiagnostic`), so the terminal cause is the LAST
- * code-bearing entry — `findLast`, not `find`: an earlier attempt's transient 503 must not classify a
- * terminal 400/auth failure as retryable. (Reverse scan rather than `findLast` — the tsconfig lib is
- * ES2022.)
- */
-function messageSignal(message: AssistantMessage): { status?: number; code?: unknown } {
-  const diagnostics = message.diagnostics ?? [];
-  for (let i = diagnostics.length - 1; i >= 0; i--) {
-    const code = diagnostics[i]?.error?.code;
-    if (code !== undefined) return { code };
-  }
-  return {};
-}
 
 /**
  * In-stream event mapping — pi events are translated ONCE into the rich `SessionEvent` vocabulary;
@@ -266,30 +183,6 @@ export interface RunControls {
  *  untrusted taps here; give third parties the read-only `events()` stream instead. */
 export type SessionObserver = (session: string, event: SessionEvent, run?: RunControls) => void;
 
-/**
- * Terminal mapping, decided by the resolved message's stopReason: pi's prompt() resolves a message
- * with stopReason "error"/"aborted" rather than throwing, so relying on catch alone would miss this
- * entire failure class (violating SPEC MUST 1).
- */
-export function toTerminal(message: AssistantMessage): AgentEvent {
-  if (message.stopReason === "aborted") {
-    // A deliberate stop (control-plane abort / harness abort), not an error — see {@link ABORTED_CODE}
-    // for the consumer contract (design §6).
-    const details = message.errorMessage ?? "run aborted";
-    return { type: "failed", details, retryable: false, code: ABORTED_CODE };
-  }
-  if (message.stopReason === "error") {
-    const details = message.errorMessage ?? `model stopped: ${message.stopReason}`;
-    return { type: "failed", details, retryable: classifyRetryable(details, messageSignal(message)) };
-  }
-  return { type: "completed" };
-}
-
-export function errorToTerminal(error: unknown): Extract<AgentEvent, { type: "failed" }> {
-  const details = error instanceof Error ? error.message : String(error);
-  return { type: "failed", details, retryable: classifyRetryable(details, errorSignal(error)) };
-}
-
 type PiHarness = Awaited<ReturnType<PiHarnessFactory>>;
 
 /** Bind the concrete pi-agent-core Session behind FastAgent's tool-runtime manager port. */
@@ -357,76 +250,6 @@ async function maybeCompact(harness: PiHarness, message: AssistantMessage): Prom
   if (!contextWindow) return;
   if (shouldCompact(calculateContextTokens(message.usage), contextWindow, DEFAULT_COMPACTION_SETTINGS)) {
     await harness.compact();
-  }
-}
-
-/**
- * Map prompt images to pi ImageContent, resizing each to model-friendly dimensions/size with pi's
- * Photon resizer (reused from pi-coding-agent, lazy-imported so the common no-image headless path never
- * loads the TUI module graph). A null resize (unresizable / Photon unavailable) keeps the original
- * bytes — the provider then applies its own limit.
- */
-export async function toPiPromptOptions(prompt: Prompt): Promise<{ images?: ImageContent[] } | undefined> {
-  if (!prompt.images || prompt.images.length === 0) return undefined;
-  const { resizeImage } = await import("@earendil-works/pi-coding-agent");
-  const images = await Promise.all(
-    prompt.images.map(async (img): Promise<ImageContent> => {
-      const resized = await resizeImage(Buffer.from(img.data, "base64"), img.mimeType, {
-        maxWidth: 1568,
-        maxHeight: 1568,
-        maxBytes: 5 * 1024 * 1024,
-      }).catch(() => null);
-      return resized
-        ? { type: "image", data: resized.data, mimeType: resized.mimeType }
-        : { type: "image", data: img.data, mimeType: img.mimeType };
-    }),
-  );
-  return { images };
-}
-
-// ── §3 EventQueue: push→pull plumbing for pi's two-port shape ────────────────
-//
-// Single-consumer async queue; single-threaded JS means no await interleaves between push and
-// drain, so no locking. Engines that are natively async-iterable would not need it. Exported for
-// the sibling L0 (`session-invoke.ts`): both engine classes have pi's two-port shape (a promise for
-// the turn + a subscription for the stream), so the fan-in is the same plumbing.
-
-export class EventQueue<T> {
-  private buffer: T[] = [];
-  private wake?: () => void;
-
-  push(item: T): void {
-    this.buffer.push(item);
-    const wake = this.wake;
-    this.wake = undefined;
-    wake?.();
-  }
-
-  /**
-   * Yield pushed events in order until `done` settles AND the buffer is drained. The terminal is
-   * produced separately (toTerminal); rejections of `done` are swallowed here (the caller awaits
-   * `run` itself) to avoid unhandled rejections.
-   */
-  async *drainUntil(done: Promise<unknown>): AsyncGenerator<T> {
-    let settled = false;
-    const onSettle = () => {
-      settled = true;
-      const wake = this.wake;
-      this.wake = undefined;
-      wake?.();
-    };
-    const finished = done.then(onSettle, onSettle);
-
-    while (true) {
-      while (this.buffer.length > 0) {
-        yield this.buffer.shift() as T;
-      }
-      if (settled) break;
-      await new Promise<void>((resolve) => {
-        this.wake = resolve;
-      });
-    }
-    await finished;
   }
 }
