+++ src/pi.ts
@@ -34,7 +34,8 @@ export {
 export type { LoadedDefinition, SkillCollision } from "./engines/pi/definition.ts";
 
 export { defineConfig, listModels, resolveModel, type FastagentConfig } from "./engines/pi/config.ts";
-export { inProcessLease, type Lease, type Release, type SessionObserver } from "./engines/pi/invoke.ts";
+export { inProcessLease, type Lease, type Release } from "./engines/pi/turn-plumbing.ts";
+export type { SessionObserver } from "./engines/pi/invoke.ts";
 export {
   createPiSessionControl,
   type CreatePiSessionControlOptions,
